+++ src/main/java/frc/robot/autochooser/chooser/AutoChooser.java
@@ -3,6 +3,7 @@ package frc.robot.autochooser.chooser;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.autochooser.event.AutoEvent;
+import frc.robot.autochooser.event.AutoEventProvider;
 
 /**
  * interface for taking in a {@link AutoEvent} and returning the corresponding {@link Command}
@@ -10,7 +11,7 @@ import frc.robot.autochooser.event.AutoEvent;
 public interface AutoChooser {
     /**
      * @return Command that corresponds to the selected {@link AutoEvent}
-     * from the {@link frc.robot.autochooser.event.AutoEventProvider}
+     * from the {@link AutoEventProvider}
      */
     Command getAutoCommand();
     Pose2d getStartingPosition();
