+++ src/main/java/frc/robot/utils/shuffleboard/SmartShuffleboard.java
@@ -8,6 +8,7 @@ import edu.wpi.first.networktables.NetworkTableValue;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardLayout;
 import edu.wpi.first.wpilibj.shuffleboard.SimpleWidget;
 import edu.wpi.first.wpilibj2.command.Command;
+
 import java.util.HashMap;
 import java.util.Map;
